+++ ukcp_updater/__init__.py
@@ -3,4 +3,4 @@ UKCP Updater
 Chris Parkinson (@chssn)
 """
 
-__version__ = "0.7.3"
+__VERSION__ = "1.0.1"