+++ ukcp_updater/__main__.py
@@ -13,8 +13,7 @@ import sys
 from loguru import logger
 
 # Local Libraries
-import ukcp_updater
-from ukcp_updater import functions
+from ukcp_updater import functions, github, scanner, __VERSION__
 
 @logger.catch
 def main():
@@ -27,28 +26,25 @@ def main():
     # Intro
     os.system("cls")
     print("-" * 60)
-    print(f"VATSIM UK Controller Pack Updater v{ukcp_updater.__version__}")
+    print(f"VATSIM UK Controller Pack Updater v{__VERSION__}")
     print("https://github.com/chssn/ukcp-updater")
     print("-" * 60)
 
     # Check that git is installed
-    git = functions.Downloader()
+    git = github.Downloader()
     if git.check_requirements():
         # Check that the repo exists, if not then clone it
         git.clone()
 
         # Check current settings
-        update = functions.CurrentInstallation()
+        update = scanner.CurrentInstallation()
         user_settings = update.user_settings()
 
         # Stash any changes and run 'git pull'
-        git.pull()
+        functions.sync_cache_to_live(git.git_path, git.live_path)
 
         # Append user settings
         update.apply_settings(user_settings, {"None": None})
-
-        # Drop the stashed files
-        git.drop_stash()
 
         # Final check before closing
         input("Update has completed. Press ENTER to close this window...")